% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-boundaries.r
\name{define_boundaries}
\alias{define_boundaries}
\title{Define the genomic region spanning all the feature types of interest.}
\usage{
define_boundaries(filtered_gtf, grouping = "gene_id")
}
\arguments{
\item{filtered_gtf}{The filtered gencode tibble}

\item{grouping}{The feature type of interest. Must be a field of the
filtered_gtf tibble (e.g., feature %in% names(filtered_gtf).
Default = "gene_id".}
}
\value{
A tibble containing the regions of interest. Variables in this
 tibble include:
 - chr - chromosome in genome build GRCh37/hg19
 - strand - DNA strand. values {+,-}
 - gene_id - gene identifier with ENSG prefix
 - gene_name - Gene name
 - agg_start - lower bound of genic unit
 - agg_end - upper bound of genic unit
 - source - comma seprated lists values of annotation sources of transcripts
   in a genic unit
 - transcript_type - comma seprated values of biotypes of transcripts in a
   genic unit
 - merge_count - number of transcripts merged in the genic unit
 - agg_size - length of the genic unit derived as a difference of agg_end
   agg_start
}
\description{
This function processes a filtered gencode tibble (such as returned from
filter_gencode()) to define the genomic region that spans all the features
of interest.
}
\details{
NOTE: for now, this works for grouping on gene_id. I may generalize later
if we want to group on other features.
}
\examples{
\dontrun{
define_boundaries(filtered_gtf, feature = "gene_id")
}
}
