% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-boundaries.r
\name{.add_ranges}
\alias{.add_ranges}
\title{Get feature ranges and add agg_start, agg_end, and merge_count variables}
\usage{
.add_ranges(reformatted_gtf)
}
\arguments{
\item{reformatted_gtf}{a filtered gtf field grouped by feature with comma-
separated lists for the source and transcript_type fields}
}
\value{
a tibble
}
\description{
Get feature ranges and add agg_start, agg_end, and merge_count variables
}
