% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-boundaries.r
\name{.parse_sources}
\alias{.parse_sources}
\title{Parse source and transcript_types columns to merge rows for transcripts with
same transcript_id. In such cases, unique sources and transcript_types
should be merged to comma-separated lists in the source and transcript_type
variables.}
\usage{
.parse_sources(grouped_gtf)
}
\arguments{
\item{grouped_gtf}{a filtered gtf field grouped by feature}
}
\value{
a tibble
}
\description{
Parse source and transcript_types columns to merge rows for transcripts with
same transcript_id. In such cases, unique sources and transcript_types
should be merged to comma-separated lists in the source and transcript_type
variables.
}
