% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{.pull_required}
\alias{.pull_required}
\title{Preliminary parsing for the attribute column of raw gtf tibble: use regular
expression to pick off the easier required fields, leave rest in "optional"}
\usage{
.pull_required(raw_gtf)
}
\arguments{
\item{raw_gtf}{A tibble from .read_gtf().}
}
\value{
A tibble with attribute column partially parsed and replaced with
the "optional" column with the remaining unparsed fields.
}
\description{
Preliminary parsing for the attribute column of raw gtf tibble: use regular
expression to pick off the easier required fields, leave rest in "optional"
}
