% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{.read_gtf}
\alias{.read_gtf}
\title{Read GENCODE gtf file to tibble. Assign fields based on tab separation
with no further parsing.}
\usage{
.read_gtf(path)
}
\arguments{
\item{path}{PATH to gencode file.}
}
\value{
A tibble
}
\description{
Read GENCODE gtf file to tibble. Assign fields based on tab separation
with no further parsing.
}
