% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{.tidy_lists}
\alias{.tidy_lists}
\title{Tidy fields with 0 to many entries: tags, ccdsids, onts.}
\usage{
.tidy_lists(messy_gtf)
}
\arguments{
\item{messy_gtf}{A tibble from .parse_optional().}
}
\value{
A tidied tibble with no list-columns.
}
\description{
Tidy fields with 0 to many entries: tags, ccdsids, onts.
}
