% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-gencode.r
\name{filter_gencode}
\alias{filter_gencode}
\title{Filter the gencode tibble by feature and tag.}
\usage{
filter_gencode(gtf, featurearg = "transcript", tagarg = "basic")
}
\arguments{
\item{gtf}{The gencode tibble}

\item{feature}{The feature type of interest. Must be one of {gene, 
transcript, exon, CDS, UTR, start_codon, stop_codon, Selenocysteine}. 
Default = "transcript"}

\item{tag}{The attribute of interest. Legal tags are listed at
https://www.gencodegenes.org/gencode_tags.html Default = "basic"}
}
\value{
A tibble containing the rows matching the feature and tag of interest
}
\description{
This function filters a gencode tibble (such as returned from 
import_gencode() to select the given feature type, taged with the given tag.
}
\examples{
\dontrun{
filter_gencode(gtf = gtf, feature = "transcript", tag = "basic")
}
}
