% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{import_gencode}
\alias{import_gencode}
\title{Import and tidy a gencode .gtf file}
\usage{
import_gencode(path)
}
\arguments{
\item{path}{Path to the gencode file.}
}
\value{
A tidied tibble
}
\description{
The gtf file format is described at
https://www.gencodegenes.org/pages/data_format.html. This function imports a
gtf and returns a tidied tibble, in which each column is a variable and each
row is an observation
}
\examples{
\dontrun{
import_gencode(path = 'path/to/file.tsv')
}
}
