% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-tag.r
\name{summarize_tag}
\alias{summarize_tag}
\title{Summarize the features in a gencode-derived tibble.}
\usage{
summarize_tag(gtf, tag = "basic")
}
\arguments{
\item{gtf}{A tibble containing gene definitions in the formate made by
genetable::import_gencode()}

\item{tag}{The tag to summarize features by - a string.}
}
\value{
A table containing the cross tabulation of features by tag.
}
\description{
summarize_tag returns a tibble containing the number of features included
in a tibble produced by genetable::import_gencode() (or similar) that are
annotated with the tag provided in the argument. Legal tags are listed at
https://www.gencodegenes.org/gencode_tags.html
}
\examples{
\dontrun{
summarize_tag(gtf = gtf, tag = "basic")
}
}
