% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{import_gencode}
\alias{import_gencode}
\title{Load a gencode file.}
\usage{
import_gencode(path)
}
\arguments{
\item{path}{Path to the gencode file.}
}
\value{
A data frame containing the rows from the file that match the 
specified feature and contain the given attribute.
}
\description{
Load a gencode file.
}
\examples{
\dontrun{
import_gencode(path = 'path/to/file.tsv')
}
}
