% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{import_gencode}
\alias{import_gencode}
\title{Get the features with attributes of interest from a gencode .gtf file}
\usage{
import_gencode(path, feature = "transcript", attribute = "tag basic")
}
\arguments{
\item{path}{Path to the gencode file.}

\item{feature}{The feature type of interest. Must be one of {gene, 
transcript, exon, CDS, UTR, start_codon, stop_codon, Selenocysteine}. 
Default = "transcript"}

\item{attribute}{The attribute of interest (from the 9th column of the gtf 
file). Default = "tag basic"}
}
\value{
A data frame containing the rows from the file that match the 
  specified feature and contain the given attribute.
}
\description{
The gtf file format is described at 
https://www.gencodegenes.org/data_format.html
}
\examples{
\dontrun{
import_gencode(path = 'path/to/file.tsv')
}
}
