% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-gencode.r
\name{import_gencode}
\alias{import_gencode}
\title{Get the features with attributes of interest from a gencode .gtf file}
\usage{
import_gencode(path, featuretag = "basic")
}
\arguments{
\item{path}{Path to the gencode file.}

\item{featuretag}{Tag to include in tag variable (defined at
https://www.gencodegenes.org/gencode_tags.html) Default = "basic"}
}
\value{
A tidied tibble
}
\description{
The gtf file format is described at 
https://www.gencodegenes.org/data_format.html
}
\examples{
\dontrun{
import_gencode(path = 'path/to/file.tsv')
}
}
